# Conflicts:
#	.github/workflows/icc/build.sh
#	MIGRATION_GUIDE.TXT
#	gcore/gdal_priv_templates.hpp
#	gcore/rasterio.cpp

+++ frmts/tiledb/tiledbdense.cpp
@@ -348,8 +348,8 @@ CPLErr TileDBRasterBand::IRasterIO(GDALRWFlag eRWFlag, int nXOff, int nYOff,
 
                     if (eErr == CE_None)
                     {
-                        CPLString osName = CPLString().Printf(
-                            "%s", CPLGetBasename(poAttrDS->GetDescription()));
+                        CPLString osName =
+                            CPLGetBasenameSafe(poAttrDS->GetDescription());
 
                         SetBuffer(poQuery.get(), eAttrType, osName, pAttrBlock,
                                   nValues);
@@ -1250,7 +1250,7 @@ CPLErr TileDBRasterDataset::TryLoadCachedXML(CSLConstList /*papszSiblingFiles*/,
         /* --------------------------------------------------------------------
          */
 
-        CPLString osPath(CPLGetPath(psPam->pszPamFilename));
+        CPLString osPath(CPLGetPathSafe(psPam->pszPamFilename));
         const CPLErr eErr = XMLInit(psTree, osPath);
 
         CPLDestroyXMLNode(psTree);
@@ -1362,7 +1362,6 @@ GDALDataset *TileDBRasterDataset::OpenInternal(GDALOpenInfo *poOpenInfo,
         poDS->nTimestamp = std::strtoull(pszTimestamp, nullptr, 10);
 
     CPLString osURI;
-    CPLString osAux;
     CPLString osSubdataset;
 
     std::string osAttrNameTmp;
@@ -1397,11 +1396,12 @@ GDALDataset *TileDBRasterDataset::OpenInternal(GDALOpenInfo *poOpenInfo,
         osURI = TileDBDataset::VSI_to_tiledb_uri(poOpenInfo->pszFilename);
     }
 
-    const char *pszArrayName = CPLGetBasename(osURI);
-    osAux.Printf("%s.tdb", pszArrayName);
+    CPLString osAux = CPLGetBasenameSafe(osURI);
+    osAux += ".tdb";
 
     // aux file is in array folder
-    poDS->SetPhysicalFilename(CPLFormFilename(osURI, osAux, nullptr));
+    poDS->SetPhysicalFilename(
+        CPLFormFilenameSafe(osURI, osAux, nullptr).c_str());
     // Initialize any PAM information.
     poDS->SetDescription(osURI);
 
@@ -1411,7 +1411,7 @@ GDALDataset *TileDBRasterDataset::OpenInternal(GDALOpenInfo *poOpenInfo,
         poDS->m_bDatasetInGroup = true;
 
         // Full resolution raster array
-        poDS->m_osArrayURI = CPLFormFilename(osURI.c_str(), "l_0", nullptr);
+        poDS->m_osArrayURI = CPLFormFilenameSafe(osURI.c_str(), "l_0", nullptr);
     }
 
     const tiledb_query_type_t eMode =
@@ -2121,7 +2121,8 @@ TileDBRasterDataset *TileDBRasterDataset::CreateLL(const char *pszFilename,
             }
 
             // Full resolution raster array
-            poDS->m_osArrayURI = CPLFormFilename(pszFilename, "l_0", nullptr);
+            poDS->m_osArrayURI =
+                CPLFormFilenameSafe(pszFilename, "l_0", nullptr);
         }
         else
         {
@@ -2167,12 +2168,11 @@ TileDBRasterDataset *TileDBRasterDataset::CreateLL(const char *pszFilename,
             }
         }
 
-        CPLString osAux;
-        const char *pszArrayName = CPLGetBasename(pszFilename);
-        osAux.Printf("%s.tdb", pszArrayName);
+        CPLString osAux = CPLGetBasenameSafe(pszFilename);
+        osAux += ".tdb";
 
         poDS->SetPhysicalFilename(
-            CPLFormFilename(pszFilename, osAux.c_str(), nullptr));
+            CPLFormFilenameSafe(pszFilename, osAux.c_str(), nullptr).c_str());
 
         // Initialize PAM information.
         poDS->SetDescription(pszFilename);
@@ -2295,13 +2295,13 @@ bool TileDBRasterDataset::DeferredCreate(bool bCreateArray)
         // be reported as subdatasets on future reads
         for (const auto &poAttrDS : m_lpoAttributeDS)
         {
-            const char *pszAttrName =
-                CPLGetBasename(poAttrDS->GetDescription());
+            const std::string osAttrName =
+                CPLGetBasenameSafe(poAttrDS->GetDescription());
             GDALRasterBand *poAttrBand = poAttrDS->GetRasterBand(1);
             int bHasNoData = false;
             const double dfNoData = poAttrBand->GetNoDataValue(&bHasNoData);
-            CreateAttribute(poAttrBand->GetRasterDataType(), pszAttrName, 1,
-                            CPL_TO_BOOL(bHasNoData), dfNoData);
+            CreateAttribute(poAttrBand->GetRasterDataType(), osAttrName.c_str(),
+                            1, CPL_TO_BOOL(bHasNoData), dfNoData);
         }
 
         if (bCreateArray)
@@ -2612,7 +2612,7 @@ TileDBRasterDataset *TileDBRasterDataset::Create(const char *pszFilename,
             {
                 aosImageStruct.SetNameValue(
                     CPLString().Printf("TILEDB_ATTRIBUTE_%i", ++i),
-                    CPLGetBasename(poAttrDS->GetDescription()));
+                    CPLGetBasenameSafe(poAttrDS->GetDescription()).c_str());
             }
         }
         poDS->SetMetadata(aosImageStruct.List(), "IMAGE_STRUCTURE");
@@ -2819,8 +2819,8 @@ void TileDBRasterDataset::LoadOverviews()
     for (int i = 0; i < m_nOverviewCountFromMetadata; ++i)
     {
         const std::string osArrayName = CPLSPrintf("l_%d", 1 + i);
-        std::string osOvrDatasetName =
-            CPLFormFilename(GetDescription(), osArrayName.c_str(), nullptr);
+        const std::string osOvrDatasetName =
+            CPLFormFilenameSafe(GetDescription(), osArrayName.c_str(), nullptr);
 
         GDALOpenInfo oOpenInfo(osOvrDatasetName.c_str(), eAccess);
         oOpenInfo.papszOpenOptions = aosOpenOptions.List();
@@ -2943,7 +2943,10 @@ CPLErr TileDBRasterDataset::IBuildOverviews(
             {
                 CPL_IGNORE_RET_VAL(poODS->Close());
                 tiledb::Array::delete_array(*m_ctx, poODS->GetDescription());
-                vfs.remove_dir(poODS->GetDescription());
+                if (vfs.is_dir(poODS->GetDescription()))
+                {
+                    vfs.remove_dir(poODS->GetDescription());
+                }
             }
             catch (const tiledb::TileDBError &e)
             {
@@ -3019,8 +3022,8 @@ CPLErr TileDBRasterDataset::IBuildOverviews(
 
             const std::string osArrayName =
                 CPLSPrintf("l_%d", 1 + int(m_apoOverviewDS.size()));
-            std::string osOvrDatasetName =
-                CPLFormFilename(GetDescription(), osArrayName.c_str(), nullptr);
+            const std::string osOvrDatasetName = CPLFormFilenameSafe(
+                GetDescription(), osArrayName.c_str(), nullptr);
 
             auto poOvrDS = std::unique_ptr<TileDBRasterDataset>(
                 Create(osOvrDatasetName.c_str(), nOXSize, nOYSize, nBands,
